TraceQL primer.
This was written alongside the Intro To MLTP repository
(https://github.com/grafana/intro-to-mltp/) where the
docker-compose.yml had been removed of the `grafana`
service, so that `npm run server` from the Pathfinder
repo could be run in replacement.

It should work, however, for any installation as the
queries use an empty selector to return all spans.

This needs more work and a dedicated data set to
carry out more complex TraceQL queries.

The intention is to copy the TraceQL breakout from
https://grafana.github.io/tracing-workshop/category/lab-2---traceql

Signed-off-by: Heds Simons <hedley.simons@grafana.com>